Add `fluent-plugin-rewrite-tag-filter` plugin
We need this so that we can edit the tags based on a log event. All
events for Dropwizard arrive with the same tags, however they have a
field we can use in order to parse/handle them differently. For example,
we want to handle HTTP logs differently from DB logs.

We should probably consider renaming this image as it no longer is just
for Sumologic.